4.3 Move wallet creation functions (`create_wallet_from_seed_phrase`, `create_wallet_from_view_key`) to `src/scanning/wallet_scanner.rs`
